cpp: trim C++ runtime (move interop to tests)
